+++ spec/requests/as_content_owner/channel_spec.rb
@@ -85,6 +85,154 @@ describe Yt::Channel, :partner do
           expect(channel.views(to: date).keys.max).to eq date.to_date
         end
       end
+
+      describe 'comments can be retrieved for a specific day' do
+        context 'in which the channel was partnered' do
+          let(:comments) { channel.comments_on 5.days.ago}
+          it { expect(comments).to be_a Float }
+        end
+
+        context 'in which the channel was not partnered' do
+          let(:comments) { channel.comments_on 20.years.ago}
+          it { expect(comments).to be_nil }
+        end
+      end
+
+      describe 'comments can be retrieved for a range of days' do
+        let(:date) { 4.days.ago }
+
+        specify 'with a given start (:since option)' do
+          expect(channel.comments(since: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:until option)' do
+          expect(channel.comments(until: date).keys.max).to eq date.to_date
+        end
+
+        specify 'with a given start (:from option)' do
+          expect(channel.comments(from: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:to option)' do
+          expect(channel.comments(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'likes can be retrieved for a specific day' do
+        context 'in which the channel was partnered' do
+          let(:likes) { channel.likes_on 5.days.ago}
+          it { expect(likes).to be_a Float }
+        end
+
+        context 'in which the channel was not partnered' do
+          let(:likes) { channel.likes_on 20.years.ago}
+          it { expect(likes).to be_nil }
+        end
+      end
+
+      describe 'likes can be retrieved for a range of days' do
+        let(:date) { 4.days.ago }
+
+        specify 'with a given start (:since option)' do
+          expect(channel.likes(since: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:until option)' do
+          expect(channel.likes(until: date).keys.max).to eq date.to_date
+        end
+
+        specify 'with a given start (:from option)' do
+          expect(channel.likes(from: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:to option)' do
+          expect(channel.likes(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'dislikes can be retrieved for a specific day' do
+        context 'in which the channel was partnered' do
+          let(:dislikes) { channel.dislikes_on 5.days.ago}
+          it { expect(dislikes).to be_a Float }
+        end
+
+        context 'in which the channel was not partnered' do
+          let(:dislikes) { channel.dislikes_on 20.years.ago}
+          it { expect(dislikes).to be_nil }
+        end
+      end
+
+      describe 'dislikes can be retrieved for a range of days' do
+        let(:date) { 4.days.ago }
+
+        specify 'with a given start (:since option)' do
+          expect(channel.dislikes(since: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:until option)' do
+          expect(channel.dislikes(until: date).keys.max).to eq date.to_date
+        end
+
+        specify 'with a given start (:from option)' do
+          expect(channel.dislikes(from: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:to option)' do
+          expect(channel.dislikes(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'shares can be retrieved for a range of days' do
+        let(:date) { 4.days.ago }
+
+        specify 'with a given start (:since option)' do
+          expect(channel.shares(since: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:until option)' do
+          expect(channel.shares(until: date).keys.max).to eq date.to_date
+        end
+
+        specify 'with a given start (:from option)' do
+          expect(channel.shares(from: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:to option)' do
+          expect(channel.shares(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'impressions can be retrieved for a specific day' do
+        context 'in which the channel was partnered' do
+          let(:impressions) { channel.impressions_on 5.days.ago}
+          it { expect(impressions).to be_a Float }
+        end
+
+        context 'in which the channel was not partnered' do
+          let(:impressions) { channel.impressions_on 20.years.ago}
+          it { expect(impressions).to be_nil }
+        end
+      end
+
+      describe 'impressions can be retrieved for a range of days' do
+        let(:date) { 4.days.ago }
+
+        specify 'with a given start (:since option)' do
+          expect(channel.impressions(since: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:until option)' do
+          expect(channel.impressions(until: date).keys.max).to eq date.to_date
+        end
+
+        specify 'with a given start (:from option)' do
+          expect(channel.impressions(from: date).keys.min).to eq date.to_date
+        end
+
+        specify 'with a given end (:to option)' do
+          expect(channel.impressions(to: date).keys.max).to eq date.to_date
+        end
+      end
     end
 
     context 'not managed by the authenticated Content Owner' do