+++ HISTORY.md
@@ -11,6 +11,7 @@ v0.7 - 2014/06/18
 * Add video.update to update title, description, tags and categoryId of a video
 * Sort channel.videos by most recent first
 * Extract Reports (earnings, views) into module with macro `has_report`
+* New channel reports: comments, likes, dislikes, shares and impressions
 
 v0.6 - 2014/06/05
 -----------------